requirements.txt
absl-py==1.0.0
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
backcall==0.2.0
brainflow==5.10.1
cachetools==5.0.0
certifi==2023.7.22
charset-normalizer==2.0.9
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.6
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
filelock==3.9.0
firebase==3.0.1
flatbuffers==2.0
fonttools==4.28.5
gast==0.5.3
gcloud==0.17.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.48.1
grpcio-tools==1.48.1
gspread==5.2.0
h5py==3.6.0
httplib2==0.20.2
idna==3.3
imageio==2.14.1
importlib-metadata==4.10.1
ipykernel==6.10.0
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.0
jupyter-core==4.9.2
jws==0.1.3
keras==2.8.0
Keras-Preprocessing==1.1.2
Kivy==2.0.0
Kivy-Garden==0.1.4
kivymd==0.104.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nptyping==1.4.4
numpy==1.22.0
oauth2client==3.0.0
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
prompt-toolkit==3.0.28
protobuf==3.20.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycryptodome==3.4.3
Pygments==2.11.1
pyparsing==3.0.6
Pyrebase==3.0.27
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2021.3
PyWavelets==1.2.0
pyzmq==22.3.0
redis==5.0.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
spotipy==2.23.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tifffile==2021.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.0.1
typish==1.9.3
urllib3==1.26.7
virtualenv==20.19.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
